Move post file helpers to helper
